requirements.txt
azure-functions
azure-storage
azure-storage-blob
azure-core